+++ index.html
@@ -254,6 +254,7 @@
   <script src="vendor/nouislider/nouislider.js"></script>
   <script src="vendor/pristine/pristine.min.js"></script>
   <script src="js/main.js" type="module"></script>
+  <script src="js/functions.js"></script>
 </body>
 
 </html>
