+++ js/fullSizePicture.js
@@ -99,4 +99,3 @@ const showBigPicture = (picture) => {
 };
 
 export { showBigPicture };
-
